+++ pybamm/input/parameters/lithium_ion/positive_electrodes/nmc_ORegan2022/nmc_LGM50_diffusivity_ORegan2022.py
@@ -1,7 +1,7 @@
 from pybamm import exp, constants
 
 
-def nmc_LGM50_diffusivity_ORegan2021(sto, T):
+def nmc_LGM50_diffusivity_ORegan2022(sto, T):
     """
     NMC diffusivity as a function of stoichiometry, in this case the
     diffusivity is taken to be a constant. The value is taken from [1].
@@ -9,9 +9,8 @@ def nmc_LGM50_diffusivity_ORegan2021(sto, T):
     References
     ----------
     .. [1] Kieran O’Regan, Ferran Brosa Planella, W. Dhammika Widanage, and Emma
-    Kendrick. "Thermal-electrochemical parametrisation of a lithium-ion battery:
-    mapping Li concentration and temperature dependencies." Journal of the
-    Electrochemical Society, submitted (2021).
+    Kendrick. "Thermal-electrochemical parameters of a high energy lithium-ion
+    cylindrical battery." Electrochimica Acta 425 (2022): 140700
 
     Parameters
     ----------