+++ pybamm/input/parameters/lithium_ion/positive_electrodes/nmc_ORegan2022/nmc_LGM50_electrolyte_exchange_current_density_ORegan2022.py
@@ -1,19 +1,18 @@
 from pybamm import exp, constants, Parameter
 
 
-def graphite_LGM50_electrolyte_exchange_current_density_ORegan2021(
+def nmc_LGM50_electrolyte_exchange_current_density_ORegan2022(
     c_e, c_s_surf, c_s_max, T
 ):
     """
-    Exchange-current density for Butler-Volmer reactions between graphite and LiPF6 in
+    Exchange-current density for Butler-Volmer reactions between NMC and LiPF6 in
     EC:DMC.
 
     References
     ----------
     .. [1] Kieran O’Regan, Ferran Brosa Planella, W. Dhammika Widanage, and Emma
-    Kendrick. "Thermal-electrochemical parametrisation of a lithium-ion battery:
-    mapping Li concentration and temperature dependencies." Journal of the
-    Electrochemical Society, submitted (2021).
+    Kendrick. "Thermal-electrochemical parameters of a high energy lithium-ion
+    cylindrical battery." Electrochimica Acta 425 (2022): 140700
 
     Parameters
     ----------
@@ -31,10 +30,9 @@ def graphite_LGM50_electrolyte_exchange_current_density_ORegan2021(
     :class:`pybamm.Symbol`
         Exchange-current density [A.m-2]
     """
-
-    i_ref = 2.668  # (A/m2)
-    alpha = 0.792
-    E_r = 4e4
+    i_ref = 5.028  # (A/m2)
+    alpha = 0.43
+    E_r = 2.401e4
     arrhenius = exp(E_r / constants.R * (1 / 298.15 - 1 / T))
 
     c_e_ref = Parameter("Typical electrolyte concentration [mol.m-3]")